requirements.txt
Flask==2.0.1
paho-mqtt==1.5.1
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
gunicorn==20.1.0
numpy==2.0.2